backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pymongo==4.6.0
motor==3.3.2
pydantic==2.5.0
pydantic-settings==2.0.3
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
fastapi-cors==0.0.6
email-validator==2.1.0
Pillow==10.1.0

# Payment processing dependencies
stripe==7.6.0
requests==2.31.0

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.24.1
pytest-mock==3.12.0
pytest-cov==4.1.0
